+++ libs/notification/src/lib/+state/index.ts
@@ -1,4 +1,2 @@
 export * from './notification.model';
-export * from './notification.query';
 export * from './notification.service';
-export * from './notification.store';
