+++ libs/notification/src/lib/notification.component.ts
@@ -1,11 +1,11 @@
 ﻿import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
-import { NotificationQuery } from './+state/notification.query';
 import { NotificationService } from './+state';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { RouterQuery } from '@datorama/akita-ng-router-store';
 import { Router } from '@angular/router';
 import { getCurrentApp, getOrgModuleAccess } from '@blockframes/utils/apps';
+import { take } from 'rxjs/operators';
 
 @Component({
   selector: 'notification-view',
@@ -15,10 +15,9 @@ import { getCurrentApp, getOrgModuleAccess } from '@blockframes/utils/apps';
 })
 export class NotificationComponent implements OnInit {
 
-  public notifications$ = this.query.selectAll();
+  public notifications$ = this.service.myNotifications$;
 
   constructor(
-    private query: NotificationQuery,
     private service: NotificationService,
     private dynTitle: DynamicTitleService,
     private router: Router,
@@ -30,8 +29,9 @@ export class NotificationComponent implements OnInit {
     this.dynTitle.setPageTitle('Notifications');
   }
 
-  markAll() {
-    for (const notification of this.query.getAll()) {
+  async markAll() {
+    const myNotifications = await this.service.myNotifications$.pipe(take(1)).toPromise();
+    for (const notification of myNotifications) {
       this.service.readNotification(notification);
     }
   }
