+++ libs/ui/src/lib/layout/marketplace/marketplace.component.ts
@@ -12,7 +12,7 @@ import { map, switchMap } from 'rxjs/operators';
 import { AuthQuery } from '@blockframes/auth/+state/auth.query';
 import { routeAnimation } from '@blockframes/utils/animations/router-animations';
 import { InvitationService } from '@blockframes/invitation/+state';
-import { NotificationQuery } from '@blockframes/notification/+state';
+import { NotificationService } from '@blockframes/notification/+state';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { MovieService, Movie } from '@blockframes/movie/+state'
 import { RouterQuery } from '@datorama/akita-ng-router-store';
@@ -28,7 +28,7 @@ import { getCurrentApp, App } from '@blockframes/utils/apps';
 export class MarketplaceComponent implements OnInit {
   public user$ = this.authQuery.select('profile');
   public wishlistCount$: Observable<number>;
-  public notificationCount$ = this.notificationQuery.selectCount();
+  public notificationCount$ = this.notificationService.myNotificationsCount$;
   public invitationCount$ = this.invitationService.myInvitations$.pipe(
     map(invitations => invitations.filter(invitation => invitation.status === 'pending').length),
   )
@@ -39,7 +39,7 @@ export class MarketplaceComponent implements OnInit {
   constructor(
     private orgService: OrganizationService,
     private invitationService: InvitationService,
-    private notificationQuery: NotificationQuery,
+    private notificationService: NotificationService,
     private authQuery: AuthQuery,
     private movieService: MovieService,
     private routerQuery: RouterQuery,
