+++ libs/utils/src/lib/routes/create-routes.ts
@@ -4,7 +4,6 @@ import { AuthGuard } from '@blockframes/auth/guard/auth.guard';
 import { EmailVerifiedGuard } from '@blockframes/auth/guard/email-verified.guard';
 import { PermissionsGuard } from '@blockframes/permissions/guard/permissions.guard';
 import { OrganizationGuard } from '@blockframes/organization/guard/organization.guard';
-import { NotificationsGuard } from '@blockframes/notification/notifications.guard';
 import { MaintenanceGuard } from '@blockframes/ui/maintenance';
 import { RequestAccessGuard } from '@blockframes/organization/guard/request-access.guard';
 
@@ -91,8 +90,7 @@ export function createRoutes({ appsRoutes, appName, landing, events }: RouteOpti
             },
             {
               path: 'o',
-              canActivate: [NotificationsGuard, PermissionsGuard, OrganizationGuard, EmailVerifiedGuard],
-              canDeactivate: [NotificationsGuard],
+              canActivate: [PermissionsGuard, OrganizationGuard, EmailVerifiedGuard],
               children
             }
           ]
